+++ src/components/ItaliaTheme/Blocks/ArgumentsInEvidence/Background.jsx
@@ -1,7 +1,12 @@
 import React from 'react';
 
 const Background = (props) => (
-  <div className="argumentInEvidence-background" {...props} />
+  <div
+    className={`argumentInEvidence-background ${
+      props.inEditMode ? 'full-width' : ''
+    }`}
+    {...props}
+  />
 );
 
 export default Background;