+++ src/components/ItaliaTheme/Blocks/Listing/SliderTemplate.jsx
@@ -3,6 +3,7 @@
 /* eslint-disable jsx-a11y/interactive-supports-focus */
 import React, { useState, useRef } from 'react';
 import PropTypes from 'prop-types';
+import cx from 'classnames';
 import { useIntl, defineMessages } from 'react-intl';
 import { Container, Row, Col } from 'design-react-kit/dist/design-react-kit';
 import Slider from 'react-slick';
@@ -34,18 +35,25 @@ const SliderTemplate = ({
   show_block_bg,
   linkTitle,
   linkHref,
+  slidesToShow = '1',
+  full_width = false,
+  show_image_title = true,
+  show_dots = true,
+  autoplay = false,
+  autoplay_speed = 2, //seconds
 }) => {
   const intl = useIntl();
   const slider = useRef(null);
-  const [autoplay, setAutoplay] = useState(false);
+  const [userAutoplay, setUserAutoplay] = useState(autoplay);
+  const nSlidesToShow = parseInt(slidesToShow);
 
   const toggleAutoplay = () => {
     if (!slider?.current) return;
-    if (autoplay) {
-      setAutoplay(false);
+    if (userAutoplay) {
+      setUserAutoplay(false);
       slider.current.slickPause();
     } else {
-      setAutoplay(true);
+      setUserAutoplay(true);
       slider.current.slickPlay();
     }
   };
@@ -69,12 +77,13 @@ const SliderTemplate = ({
   };
 
   const settings = {
-    dots: true,
+    dots: show_dots,
     infinite: true,
+    autoplay: autoplay,
     speed: 500,
-    slidesToShow: 1,
-    slidesToScroll: 1,
-    autoplaySpeed: 2000,
+    slidesToShow: nSlidesToShow,
+    slidesToScroll: nSlidesToShow,
+    autoplaySpeed: autoplay_speed * 1000,
     pauseOnHover: true,
     pauseOnFocus: true,
     pauseOnDotsHover: true,
@@ -85,12 +94,24 @@ const SliderTemplate = ({
     accessibility: true,
     nextArrow: <NextArrow />,
     prevArrow: <PrevArrow />,
+    responsive: [
+      {
+        breakpoint: 980,
+        settings: {
+          slidesToShow: 1,
+        },
+      },
+    ],
   };
 
   //const getCaption = (item) => item.description ?? item.rights ?? null;
 
   return (
-    <div className="sliderTemplate">
+    <div
+      className={cx(`sliderTemplate slidesToShow-${nSlidesToShow || 1}`, {
+        'no-margin': full_width,
+      })}
+    >
       <Container className="px-4">
         {title && (
           <Row>
@@ -99,28 +120,39 @@ const SliderTemplate = ({
             </Col>
           </Row>
         )}
-        <div className="slider-container px-4 px-md-0">
+        <div
+          className={cx('slider-container', {
+            'px-4 px-md-0': !full_width,
+            'full-width': full_width,
+          })}
+        >
           <div className="it-carousel-all it-card-bg">
-            <div className="play-pause-wrapper">
-              <button
-                onClick={() => toggleAutoplay()}
-                title={
-                  autoplay
-                    ? intl.formatMessage(messages.pause)
-                    : intl.formatMessage(messages.play)
-                }
-              >
-                <Icon icon={autoplay ? 'pause' : 'play'} />
-                <span>{autoplay ? 'pause' : 'play'}</span>
-              </button>
-            </div>
+            {items?.length > nSlidesToShow && (
+              <div className="play-pause-wrapper">
+                <button
+                  onClick={() => toggleAutoplay()}
+                  title={
+                    userAutoplay
+                      ? intl.formatMessage(messages.pause)
+                      : intl.formatMessage(messages.play)
+                  }
+                >
+                  <Icon icon={userAutoplay ? 'pause' : 'play'} />
+                  <span>{userAutoplay ? 'pause' : 'play'}</span>
+                </button>
+              </div>
+            )}
+
             <Slider {...settings} ref={slider}>
-              {items.map((item) => {
+              {items.map((item, index) => {
                 const image =
                   item.image || item.immagine_testata || item.foto_persona;
 
                 return (
-                  <div className="it-single-slide-wrapper" key={item['@id']}>
+                  <div
+                    className="it-single-slide-wrapper"
+                    key={item['@id'] + index}
+                  >
                     <div className="slide-wrapper">
                       <figure className="img-wrapper">
                         {image && (
@@ -137,14 +169,16 @@ const SliderTemplate = ({
                         <figcaption>{getCaption(item)}</figcaption>
                       )} */}
                       </figure>
-                      <UniversalLink
-                        item={item}
-                        title={intl.formatMessage(messages.viewImage)}
-                      >
-                        <div className="slide-title">
-                          {item.title} <Icon icon="arrow-right" />
-                        </div>
-                      </UniversalLink>
+                      {show_image_title && (
+                        <UniversalLink
+                          item={item}
+                          title={intl.formatMessage(messages.viewImage)}
+                        >
+                          <div className="slide-title">
+                            {item.title} <Icon icon="arrow-right" />
+                          </div>
+                        </UniversalLink>
+                      )}
                     </div>
                   </div>
                 );