+++ src/components/ItaliaTheme/View/FAQ/FaqFolder/FaqFolderView.jsx
@@ -13,6 +13,7 @@ import {
   PageHeader,
   RelatedItems,
   FaqFolderPlaceholderAfterContent,
+  FaqFolderPlaceholderAfterRelatedItems,
   TextOrBlocks,
   RelatedItemInEvidence,
   PageMetadata,
@@ -20,8 +21,10 @@ import {
   FaqFolderTree,
 } from '@italia/components/ItaliaTheme/View';
 import { useDebouncedEffect } from '@italia/helpers';
-import { getContent, resetContent } from '@plone/volto/actions';
+import { resetContent } from '@plone/volto/actions';
 import { flattenToAppURL } from '@plone/volto/helpers';
+
+import { GET_CONTENT } from '@plone/volto/constants/ActionTypes';
 
 const messages = defineMessages({
   no_results: {
@@ -55,7 +58,15 @@ const FaqFolderView = ({ content }) => {
       structure_url +
       (searchableText ? '?SearchableText=' + searchableText + '*' : '');
 
-    dispatch(getContent(flattenToAppURL(url), null, FAQ_FOLDER_KEY));
+    const dispatch_obj = {
+      type: GET_CONTENT,
+      subrequest: FAQ_FOLDER_KEY,
+      request: {
+        op: 'get',
+        path: `${flattenToAppURL(url)}`,
+      },
+    };
+    dispatch(dispatch_obj);
 
     return () => dispatch(resetContent(FAQ_FOLDER_KEY));
   };
@@ -120,8 +131,9 @@ const FaqFolderView = ({ content }) => {
       <FaqFolderPlaceholderAfterContent content={content} />
 
       <RelatedItems content={content} list={content?.servizi_collegati ?? []} />
-
       <RelatedItemInEvidence content={content} />
+
+      <FaqFolderPlaceholderAfterRelatedItems content={content} />
     </>
   );
 };